Move showcase functions to separate file
